feat: Add GSAP animation library
This commit adds the GSAP (GreenSock Animation Platform) library to the
project. This library will be used for creating animations and
transitions.

The package-lock.json and package.json files are updated to include GSAP
as a dependency.

A basic animation is implemented in the HeroSection component to
demonstrate the library's functionality. This animation scales the hero
image up and down repeatedly.